refactor: update import paths and component names in UI files
