+++ src/app/projects/page.tsx
@@ -1,6 +1,6 @@
-import fetchProjects from './lib/fetch-projects.lib';
-import HeadingNavigation from './ui/heading-navigation.component';
-import ProjectList from './ui/project-list.component';
+import fetchProjects from './lib/fetch-projects';
+import HeadingNavigation from './ui/heading-navigation';
+import ProjectList from './ui/project-list';
 
 export default async function Portfolio() {
   const projects = await fetchProjects();
